Fix compiler warning "redundant cast"
Former-commit-id: c5dc25e99edbad503ed5046dd0c544d9242479a1

+++ src/mediathek/tool/TModel.java
@@ -44,7 +44,7 @@ public class TModel extends DefaultTableModel {
         // die Indexspalte ist die SPALTE 0!!!!
         int ret = 0;
         for (List<Integer> list : (Iterable<List<Integer>>) getDataVector()) {
-            if ((Integer) list.get(0) == idxWert) {
+            if (list.get(0) == idxWert) {
                 return ret;
             }
             ++ret;
